requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
mysql-connector-python==9.5.0
pyTelegramBotAPI==4.29.1
requests==2.32.5
urllib3==2.6.2
